refactor(search_playground): abstract page template to shared component
